requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart
aiosqlite
faster-whisper>=1.0.0
sentence-transformers>=2.6.0
chromadb>=0.4.22
pymupdf>=1.23.0
google-generativeai>=0.5.0
Pillow>=10.0.0
aiofiles
sse-starlette
